# swagger.yaml
swagger: '2.0'
info:
  description: Набор сервисов для учета на предприятии /ООО Ромашка/
  version: "1.0.0"
  title: RestApi

produces:
  - application/json
  - application/xml
  - application/text

basePath: "/api"

paths:
  /report/formats:
    get:
      summary: Получить список форматов для отчетов
      operationId: main.formats
      produces:
        - application/json
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка     

  /report/models:
    get:
      summary: Получить список моделей
      operationId: main.models
      produces:
        - application/json
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка 

  /type_filter:
    get:
      summary: Получить список типов фильтрации
      operationId: main.get_type_filter
      produces:
        - application/json
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка 

  /report/range/{format}:
    get:
      summary: Получить отчет по единицам измерения
      operationId: main.report_range
      produces:
        - application/text
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка
    parameters:
    - name: format
      in: path
      description: Уникальный код формата
      required: true
      type: string 
    
  /report/group/{format}:
    get:
      summary: Получить отчет по группам номенклатур
      operationId: main.report_group
      produces:
        - application/text
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка
    parameters:
    - name: format
      in: path
      description: Уникальный код формата
      required: true
      type: string

  /report/nomenclature/{format}:
    get:
      summary: Получить отчет по номенклатурам
      operationId: main.report_nomenclature
      produces:
        - application/text
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка
    parameters:
    - name: format
      in: path
      description: Уникальный код формата
      required: true
      type: string

  /report/recipe/{format}:
    get:
      summary: Получить отчет по рецептам
      operationId: main.report_recipe
      produces:
        - application/text
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка
    parameters:
    - name: format
      in: path
      description: Уникальный код формата
      required: true
      type: string

  /report/storage/{format}:
    get:
      summary: Получить отчет по рецептам
      operationId: main.report_storage
      produces:
        - application/text
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка
    parameters:
    - name: format
      in: path
      description: Уникальный код формата
      required: true
      type: string

  /report/transaction/{format}:
    get:
      summary: Получить отчет по транзакциям
      operationId: main.report_transaction
      produces:
        - application/text
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка
    parameters:
    - name: format
      in: path
      description: Уникальный код формата
      required: true
      type: string
  
  /report/turnover/{format}:
    get:
      summary: Получить отчет по оборотам
      operationId: main.report_turnover
      produces:
        - application/text
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка
    parameters:
    - name: format
      in: path
      description: Уникальный код формата
      required: true
      type: string

  /filter/{model}:
    post:
      summary: Фильтрация данных по модели
      operationId: main.filter_data
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: model
          in: path
          required: true
          type: string
          description: Модели для фильтрации. Смотреть существующие модели /api/report/models.
        - in: body
          name: filter
          description: Фильтр. Типы фильтрации можно посмотреть /api/type_filter.
          required: true
          schema:
            type: object
            properties:
              name:
                type: string
                default: ""
              type_filter_name:
                type: string
                default: "EQUALE"
              id:
                type: string
                default: ""
              type_filter_id:
                type: string
                default: "EQUALE"
      responses:
        200:
          description: OK
          schema:
            type: object
        400:
          description: Данная модель не найдена. Смотреть существующие модели /api/report/models
          schema:
            type: object
        404:
          description: Данных для фильтрации по модели нет.
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка

  /transaction/filter:
    post:
      summary: Фильтрация транзакций по складу или номенклатуре
      operationId: main.filter_transaction
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: filter
          description: Фильтрация по storage или nomenclature. Типы фильтрации можно посмотреть /api/type_filter.
          required: true
          schema:
            type: object
            properties:
              model:
                type: string
                default: ""
              name:
                type: string
                default: ""
              type_filter_name:
                type: string
                default: "EQUALE"
              id:
                type: string
                default: ""
              type_filter_id:
                type: string
                default: "EQUALE"
      responses:
        200:
          description: OK
          schema:
            type: object
        400:
          description: Данная модель не найдена. Возможна фильтрация по storage или nomenclature
          schema:
            type: object
        404:
          description: Данных для фильтрации по модели нет.
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка

  /turnover/filter:
    post:
      summary: Фильтрация складских оборотов по складу, номенклатуре или периоду
      operationId: main.filter_turnover
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: filter
          description: Фильтрация по storage, nomenclature или периоду. Типы фильтрации можно посмотреть /api/type_filter.
          required: true
          schema:
            type: object
            properties:
              model:
                type: string
                default: ""
              name:
                type: string
                default: ""
              type_filter_name:
                type: string
                default: "EQUALE"
              id:
                type: string
                default: ""
              type_filter_id:
                type: string
                default: "EQUALE"
              start_period:
                type: string
                format: date-time
                default: "2024-01-01T00:00:00Z"
              end_period:
                type: string
                format: date-time
                default: "2024-12-31T23:59:59Z"
      responses:
        200:
          description: OK
          schema:
            type: object
        400:
          description: Данная модель не найдена. Возможна фильтрация по storage или nomenclature
          schema:
            type: object
        404:
          description: Данных для фильтрации по модели нет.
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка